requirements.txt
# crewai==1.14.2
# crewai-tools==1.14.2
# langchain-openai==1.1.14
# pydantic==2.11.10
# pydantic-settings==2.10.1
# pydantic_core==2.33.2
# PyGithub==2.9.1
# python-dotenv==1.1.1
# requests==2.33.1
# requests-oauthlib==2.0.0
# requests-toolbelt==1.0.0
# rich==14.3.4
# setuptools==82.0.1

crewai==1.14.2
crewai-tools==1.14.2
google-genai==1.73.1
langchain-core==1.3.0
langchain-google-genai==4.2.2
langchain-openai==1.1.14
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyGithub==2.9.1
python-dotenv==1.1.1
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.4
setuptools==82.0.1